move key_signals to src/utility
