+++ cockatrice/CMakeLists.txt
@@ -7,7 +7,6 @@ project(Cockatrice VERSION "${PROJECT_VERSION_MAJOR}.${PROJECT_VERSION_MINOR}.${
 set(cockatrice_SOURCES
     ${VERSION_STRING_CPP}
     # sort by alphabetical order, so that there is no debate about where to add new sources to the list
-    src/client/game_logic/key_signals.cpp
     src/client/get_text_with_max.cpp
     src/client/menus/deck_editor/deck_editor_menu.cpp
     src/client/network/client_update_checker.cpp
@@ -256,6 +255,7 @@ set(cockatrice_SOURCES
     src/settings/shortcut_treeview.cpp
     src/settings/shortcuts_settings.cpp
     src/utility/card_info_comparator.cpp
+    src/utility/key_signals.cpp
     src/utility/levenshtein.cpp
     src/utility/logger.cpp
     src/utility/sequence_edit.cpp
