+++ cockatrice/src/client/ui/widgets/deck_editor/deck_editor_database_display_widget.h
@@ -3,7 +3,7 @@
 
 #include "../../../../deck/custom_line_edit.h"
 #include "../../../../game/cards/card_database_model.h"
-#include "../../../game_logic/key_signals.h"
+#include "../../../../utility/key_signals.h"
 #include "../../../tabs/abstract_tab_deck_editor.h"
 
 #include <QHBoxLayout>
