+++ cockatrice/src/client/ui/widgets/deck_editor/deck_editor_filter_dock_widget.h
@@ -1,7 +1,7 @@
 #ifndef DECK_EDITOR_FILTER_DOCK_WIDGET_H
 #define DECK_EDITOR_FILTER_DOCK_WIDGET_H
 
-#include "../../../game_logic/key_signals.h"
+#include "../../../../utility/key_signals.h"
 #include "../../../tabs/abstract_tab_deck_editor.h"
 
 #include <QDockWidget>
