+++ cockatrice/src/utility/key_signals.cpp
