+++ cockatrice/src/utility/key_signals.h
